game_agent/config.py
# config.py
import os
from dotenv import load_dotenv, find_dotenv
from openai import AsyncOpenAI
from game_agents import Agent, Runner, OpenAIChatCompletionsModel

# Load .env
load_dotenv(find_dotenv())

api_key = os.getenv("OPENAI_API_KEY")

# Check API key
if not api_key:
    raise ValueError("OPENAI_API_KEY is not set. Please ensure it is defined in your .env file.")

# ✅ Correct OpenAI Client
external_client = AsyncOpenAI(
    api_key=api_key,
    base_url="https://api.openai.com/v1"  # ✅ OpenAI base URL
)

# ✅ Correct model name
model_name = "gpt-4o"  # or "gpt-4o-mini"

# ✅ Wrap model into OpenAIChatCompletionsModel so imports work
model = OpenAIChatCompletionsModel(
    model=model_name,
    openai_client=external_client
)

# Game Configuration
PLAYER_STARTING_HP = 10
MONSTER_STARTING_HP = 5
PLAYER_ATTACK_DAMAGE = 2
MONSTER_ATTACK_DAMAGE = 3
PLAYER_HIT_THRESHOLD = 8
MONSTER_HIT_THRESHOLD = 10

# Game State
STARTING_LOCATION = "Whispering Woods"
STARTING_INVENTORY = []

# Agent Configuration
AGENT_TEMPERATURE = 0.7

# Combat Configuration
DICE_SIDES = 20

# OpenAI Configuration
OPENAI_API_KEY = api_key
DEFAULT_MODEL = model_name
