game_agent/requirements.txt
chainlit
openai
python-dotenv

